Moving exception to it's own file
