+++ src/bunit.web/TestDoubles/HttpClient/PlaceholderHttpMessageHandler.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Net.Http;
+﻿using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -12,7 +11,7 @@ namespace Bunit.TestDoubles.HttpClient
 	public class PlaceholderHttpMessageHandler : HttpMessageHandler
 	{
 		/// <summary>
-		/// 
+		///
 		/// </summary>
 		/// <param name="request"></param>
 		/// <param name="cancellationToken"></param>
@@ -21,27 +20,6 @@ namespace Bunit.TestDoubles.HttpClient
 		protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
 		{
 			throw new MissingMockHttpClientException(request);
-		}
-	}
-
-	/// <summary>
-	/// Exception use to indicate that a mock HttpClient is required by a test
-	/// but was not provided.
-	/// </summary>
-	public class MissingMockHttpClientException : Exception
-	{
-		/// <summary>
-		/// The request that was sent via the http client
-		/// </summary>
-		public HttpRequestMessage Request { get; }
-
-		/// <summary>
-		///
-		/// </summary>
-		public MissingMockHttpClientException(HttpRequestMessage request)
-			: base($"This test requires a HttpClient to be supplied, because the component under test invokes the HttpClient during the test. The request that was sent is contained within the '{nameof(Request)}' attribute of this exception. Guidance on mocking the HttpClient is available in the testing library's Wiki.")
-		{
-			Request = request;
 		}
 	}
 }
